+++ src/resolvers/projectResolver.ts
@@ -1,9 +1,11 @@
-import { Arg, Mutation, Query, Resolver } from "type-graphql";
+import { Arg, Ctx, Mutation, Query, Resolver } from "type-graphql";
 import { iProject } from "../interfaces/InputType";
 import { Project } from "../models/project.model";
 import projectService from "../services/projectService";
 import string from "string-sanitizer";
 import { fileManager } from "../tools/fileManager";
+import { Context } from "apollo-server-core";
+import { TokenPayload } from "../tools/createApolloServer";
 
 @Resolver(iProject)
 export class ProjectResolver {
@@ -12,9 +14,11 @@ export class ProjectResolver {
     @Arg("userId") userId: number,
     @Arg("name") name: string,
     @Arg("description") description: string,
-    @Arg("isPublic") isPublic: boolean
+    @Arg("isPublic") isPublic: boolean,
+    @Ctx() ctx: Context<TokenPayload>
   ): Promise<Project> {
     try {
+      if (userId !== ctx.id) throw new Error("userId not allowed");
       // On Stock un timestamp pour avoir un nom unique
       const timeStamp = Date.now();
       // On supprime les espaces et les caractères spéciaux du nom du projet
@@ -40,9 +44,38 @@ export class ProjectResolver {
   }
 
   @Query(() => [Project])
-  async getAllProjects(): Promise<Project[]> {
+  async getPublicProjects(
+    @Ctx() ctx: Context<TokenPayload>
+  ): Promise<Project[]> {
     try {
-      return await projectService.getAll();
+      const projects = await projectService.getAll();
+      return projects.filter((project) => project.isPublic);
+    } catch (err) {
+      console.error(err);
+      throw new Error("Can't find public Projects");
+    }
+  }
+
+  @Query(() => [Project])
+  async getSharedWithMeProjects(
+    @Ctx() ctx: Context<TokenPayload>
+  ): Promise<Project[]> {
+    try {
+      const projects = await projectService.getAll();
+      return projects.filter((project) =>
+        project.projectShare.map((pshare) => pshare.userId).includes(ctx.id)
+      );
+    } catch (err) {
+      console.error(err);
+      throw new Error("Can't find shared with me Projects");
+    }
+  }
+
+  @Query(() => [Project])
+  async getAllProjects(@Ctx() ctx: Context<TokenPayload>): Promise<Project[]> {
+    try {
+      const projects = await projectService.getAll();
+      return projects.filter((project) => project.userId === ctx.id);
     } catch (err) {
       console.error(err);
       throw new Error("Can't find all Projects");
@@ -50,9 +83,13 @@ export class ProjectResolver {
   }
 
   @Query(() => Project)
-  async getProjectById(@Arg("projectId") projectId: number): Promise<Project> {
+  async getProjectById(
+    @Arg("projectId") projectId: number,
+    @Ctx() ctx: Context<TokenPayload>
+  ): Promise<Project[]> {
     try {
-      return await projectService.getById(projectId);
+      const projects = await projectService.getById(projectId);
+      return projects.filter((project) => project.userId === ctx.id);
     } catch (err) {
       console.error(err);
       throw new Error("Can't find Project");
@@ -62,9 +99,12 @@ export class ProjectResolver {
   @Mutation(() => Project)
   async updateProject(
     @Arg("project") project: iProject,
-    @Arg("projectId") projectId: number
-  ): Promise<Project> {
+    @Arg("projectId") projectId: number,
+    @Ctx() ctx: Context<TokenPayload>
+  ): Promise<Project[]> {
     try {
+      const [_project] = await projectService.getById(projectId);
+      if (_project.userId !== ctx.id) throw new Error("userId not allowed");
       return await projectService.update(project, projectId);
     } catch (err) {
       console.error(err);
@@ -73,10 +113,15 @@ export class ProjectResolver {
   }
 
   @Mutation(() => Project)
-  async deleteProject(@Arg("projectId") projectId: number): Promise<Project> {
+  async deleteProject(
+    @Arg("projectId") projectId: number,
+    @Ctx() ctx: Context<TokenPayload>
+  ): Promise<Project> {
     try {
-      const project = await projectService.getById(projectId);
+      const [project] = await projectService.getById(projectId);
       // Suppression du dossier du projet sur le server
+
+      if (project.userId !== ctx.id) throw new Error("userId not allowed");
 
       await fileManager.deleteProjectFolder(project.id_storage_number);
       await projectService.delete(projectId);
@@ -94,11 +139,15 @@ export class ProjectResolver {
   async createSubFolder(
     @Arg("projectId") projectId: number,
     @Arg("subFolderName") subFolderName: string,
-    @Arg("clientPath") clientPath: string
+    @Arg("clientPath") clientPath: string,
+    @Ctx() ctx: Context<TokenPayload>
   ): Promise<Project | undefined> {
     try {
       // On stock les informations du projet dans une variable
-      const project: Project = await projectService.getById(projectId);
+      const project: any = await projectService.getById(projectId);
+
+      if (project[0].user.id !== ctx.id) throw new Error("userId not allowed");
+
       return await fileManager.createOneSubFolder(
         project,
         clientPath,
