+++ src/resolvers/userResolver.ts
@@ -1,18 +1,22 @@
-import { Arg, Mutation, Query, Resolver } from "type-graphql";
+import { Context } from "apollo-server-core";
+import { Arg, Mutation, Query, Resolver, Ctx } from "type-graphql";
 import { iUser } from "../interfaces/InputType";
 import { User } from "../models/user.model";
 import authService from "../services/authService";
 import userService from "../services/userService";
+import { TokenPayload } from "../tools/createApolloServer";
 
 @Resolver(iUser)
 export class UserResolver {
   @Mutation(() => User)
   async createUser(
     @Arg("email") email: string,
-    @Arg("password") password: string
+    @Arg("password") password: string,
+    @Arg("login") login: string,
+    @Ctx() ctx: Context<TokenPayload>
   ): Promise<User> {
     try {
-      const userFromDB = await userService.create(email, password);
+      const userFromDB = await userService.create(email, password, login);
       return userFromDB;
     } catch (err) {
       console.error(err);
@@ -21,19 +25,24 @@ export class UserResolver {
   }
 
   @Query(() => [User])
-  async getAllUsers(): Promise<User[]> {
+  async getAllUsers(@Ctx() ctx: Context<TokenPayload>): Promise<User[]> {
     try {
-      return await userService.getAll();
+      const users = await userService.getAll();
+      return users.filter((user) => user.id === ctx.id);
     } catch (err) {
       console.error(err);
       throw new Error("Can't get all Users");
     }
   }
 
-  @Query(() => User)
-  async getUserById(@Arg("userId") userId: number): Promise<User> {
+  @Query(() => [User])
+  async getUserById(
+    @Arg("userId") userId: number,
+    @Ctx() ctx: Context<TokenPayload>
+  ): Promise<User[]> {
     try {
-      return await userService.getById(userId);
+      if (userId === ctx.id) return await userService.getById(userId);
+      else throw new Error("id not allowed");
     } catch (err) {
       console.error(err);
       throw new Error("Can't get User");
@@ -43,7 +52,8 @@ export class UserResolver {
   @Query(() => String)
   async getToken(
     @Arg("email") email: string,
-    @Arg("password") password: string
+    @Arg("password") password: string,
+    @Ctx() ctx: Context<TokenPayload>
   ): Promise<String> {
     try {
       // Récupérer l'utilisateur dans la bdd suivant l'email
@@ -54,6 +64,7 @@ export class UserResolver {
       ) {
         // Créer un nouveau token => signer un token
         const token = authService.signJwt({
+          id: userFromDB.id,
           email: userFromDB.email,
         });
 
@@ -71,10 +82,12 @@ export class UserResolver {
   @Mutation(() => User)
   async updateUser(
     @Arg("User") User: iUser,
-    @Arg("userId") userId: number
-  ): Promise<User> {
+    @Arg("userId") userId: number,
+    @Ctx() ctx: Context<TokenPayload>
+  ): Promise<User[]> {
     try {
-      return await userService.update(User, userId);
+      if (userId === ctx.id) return await userService.update(User, userId);
+      else throw new Error("id not allowed");
     } catch (err) {
       console.error(err);
       throw new Error("Can't update User");
@@ -82,9 +95,13 @@ export class UserResolver {
   }
 
   @Mutation(() => User)
-  async deleteUser(@Arg("userId") userId: number): Promise<User> {
+  async deleteUser(
+    @Arg("userId") userId: number,
+    @Ctx() ctx: Context<TokenPayload>
+  ): Promise<User[]> {
     try {
-      return await userService.delete(userId);
+      if (userId === ctx.id) return await userService.delete(userId);
+      else throw new Error("id not allowed");
     } catch (err) {
       console.error(err);
       throw new Error("Can't delete User");
